refactor: extract data models into a dedicated module and update references across the codebase
