+++ src/config/artifact.rs
@@ -1,3 +1,4 @@
+use crate::model::artifact::Artifact;
 use serde::{Deserialize, Serialize};
 use serde_yaml;
 use std::fs;
@@ -8,23 +9,6 @@ use super::error::ConfigError;
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ArtifactsPlanConfig {
     pub artifacts: Vec<Artifact>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Artifact {
-    pub name: String,
-    pub module: String,
-    pub role: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub inputs: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub outputs: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<Vec<String>>,
 }
 
 impl ArtifactsPlanConfig {
