+++ src/config/placement.rs
@@ -1,3 +1,4 @@
+use crate::model::placement::RolePlacement;
 use serde::{Deserialize, Serialize};
 use serde_yaml;
 use std::collections::HashMap;
@@ -9,23 +10,6 @@ use super::error::ConfigError;
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct PlacementRulesConfig {
     pub roles: HashMap<String, RolePlacement>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SidecarPlacement {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub contract_dir: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub prompt_dir: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RolePlacement {
-    pub path: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub file_extension: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sidecar: Option<SidecarPlacement>,
 }
 
 impl PlacementRulesConfig {
