+++ src/config/project.rs
@@ -1,3 +1,4 @@
+use crate::model::project::{Module, Project, Workspace};
 use serde::{Deserialize, Serialize};
 use serde_yaml;
 use std::fs;
@@ -13,27 +14,6 @@ pub struct ProjectConfig {
     pub modules: Vec<Module>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub metadata: Option<serde_yaml::Value>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Project {
-    pub name: String,
-    pub architecture_style: String,
-    pub language: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Workspace {
-    pub enabled: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub members: Option<Vec<String>>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Module {
-    pub name: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub features: Option<Vec<String>>,
 }
 
 impl ProjectConfig {
