+++ src/model/mod.rs
@@ -1,0 +1,3 @@
+pub mod artifact;
+pub mod placement;
+pub mod project;
